Refactor manager classes into Infrastructure.Managers
