+++ src/Infrastructure/Managers/MedicineRecordManager.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 Team6. All rights reserved. 
 //  No warranty, explicit or implicit, provided.
 
-namespace Infrastructure.Services;
+namespace Infrastructure.Managers;
 
 public class MedicineRecordManager(HttpClient httpClient)
 {
