+++ src/Infrastructure/Managers/MedicineStatusManager.cs
@@ -6,7 +6,7 @@ using System.Net.Http.Json;
 using Core.DTOs;
 using Core.Interfaces.Managers;
 
-namespace Infrastructure.Services;
+namespace Infrastructure.Managers;
 /// <summary>
 /// Provides HTTP-based access to medicine and painkiller status from the API.
 /// </summary>
